sandbox's config lives in cmd/config
